move serviceIndicator
