+++ serviceapi/serviceapi.go
@@ -515,55 +515,6 @@ func serviceSearchMatch(rw http.ResponseWriter, req *http.Request) {
 	fmt.Fprintf(rw, string(buf))
 }
 
-// Process a new indicator request
-func serviceIndicator(rw http.ResponseWriter, req *http.Request) {
-	req.ParseMultipartForm(10000000)
-
-	val := req.FormValue("params")
-	if val == "" {
-		logf("no indicator criteria specified")
-		http.Error(rw, "no indicator criteria specified", 500)
-		return
-	}
-	var params slib.IndicatorParams
-	err := json.Unmarshal([]byte(val), &params)
-	if err != nil {
-		logf(err.Error())
-		http.Error(rw, err.Error(), 500)
-		return
-	}
-
-	op := opContext{}
-	op.newContext(dbconn, false, req.RemoteAddr)
-
-	for _, x := range params.Indicators {
-		if x.Host == "" {
-			continue
-		}
-		err = updateDynamicHost(op, x.Host, "indicator request", 90)
-		if err != nil {
-			op.logf(err.Error())
-			http.Error(rw, err.Error(), 500)
-			return
-		}
-		err = processIndicator(op, x)
-		if err != nil {
-			op.logf(err.Error())
-			http.Error(rw, err.Error(), 500)
-			return
-		}
-	}
-
-	ir := slib.IndicatorResponse{OK: true}
-	buf, err := json.Marshal(&ir)
-	if err != nil {
-		op.logf(err.Error())
-		http.Error(rw, err.Error(), 500)
-		return
-	}
-	fmt.Fprint(rw, string(buf))
-}
-
 // Periodically prune old RRAs
 func rraCleanup() {
 	defer func() {
